refactor to remove clutter from 'deploy' command
